simplify clientRouter file
